ref #5 - some refactoring after investigation
there is a timing bug when the plugin is called multiple times (in
the order of tens of thousands of times). having dug a bit the slow
down is related to calling DB::enable_profile and only because
we are passing in a custom file name, i.e. something low level in
Devel::NYTProf itself. if we comment out that line or remove the
file name argument from the call to DB::enable_profile there is
no slow down.

there's nothing we can do here. we could perhaps raise a bug/test
case with Devel::NYTProf but given the use case for this bug i
doubt it will be a priority. looking at the source for Devel::NYTProf
i see this block[1] gets executed when a file is passed into
DB::enable_profile so the bug is somewhere around that chunk of
code and possibly in the close_output_file[2] function

(thinking out loud here: is it possible that we are cumulatively
adding to the original file, and each time the plugin runs and we
change the filename we are adding a bit more to it? I'm not sure
that's the bug given we give the start=no argument to Devel::NYTProf,
so it *shouldn't* be profiling anything *until* we call
DB::enable_profile)

having done some investigation here noticed that the file argument
in the startup is redundant, so remove it. can also refactor out a
few lines in prior to the DB::finish_profile call

[1] https://github.com/timbunce/devel-nytprof/blob/a520cac8949d94b9301df974822ccdbb32104d7a/NYTProf.xs#L2991

[2] https://github.com/timbunce/devel-nytprof/blob/a520cac8949d94b9301df974822ccdbb32104d7a/NYTProf.xs#L1869